+++ internal/Picker/Picker.js
@@ -67,7 +67,6 @@ const selectDecIcon = selectIcon('decrementIcon', 'triangledown', 'triangleleft'
 const forwardBlur = forward('onBlur'),
 	forwardFocus = forward('onFocus'),
 	forwardKeyDown = forward('onKeyDown'),
-	forwardMouseDown = forward('onMouseDown'),
 	forwardKeyUp = forward('onKeyUp'),
 	forwardWheel = forward('onWheel');
 
@@ -559,15 +558,6 @@ const PickerBase = class extends React.Component {
 
 	emulateMouseUp = new Job(this.clearPressedState, 175);
 
-	handleMouseDown = (ev) => {
-		const {joined, orientation} = this.props;
-		forwardMouseDown(ev, this.props);
-
-		if (joined && orientation === 'horizontal') {
-			this.setIncPickerButtonPressed();
-		}
-	};
-
 	handleUp = () => {
 		if (this.props.joined && (this.pickerButtonPressed !== 0 || this.state.pressed !== 0)) {
 			this.emulateMouseUp.start();
@@ -576,6 +566,10 @@ const PickerBase = class extends React.Component {
 
 	handleDown = () => {
 		const {joined, orientation} = this.props;
+
+		if (joined && orientation === 'horizontal') {
+			this.setIncPickerButtonPressed();
+		}
 
 		if (joined && this.pickerButtonPressed === 1) {
 			this.handleIncrement();
@@ -964,7 +958,6 @@ const PickerBase = class extends React.Component {
 				onKeyDown={this.handleKeyDown}
 				onKeyUp={this.handleKeyUp}
 				onUp={this.handleUp}
-				onMouseDown={this.handleMouseDown}
 				onMouseLeave={this.clearPressedState}
 				orientation={orientation}
 				ref={this.initContainerRef}