+++ Input/Input.js
@@ -2,11 +2,13 @@ import {handle, adaptEvent, forKey, forward} from '@enact/core/handle';
 import kind from '@enact/core/kind';
 import {extractAriaProps} from '@enact/core/util';
 import Spotlight from '@enact/spotlight';
+import {spotlightDefaultClass} from '@enact/spotlight/SpotlightContainerDecorator';
 import {useAnnounce} from '@enact/ui/AnnounceDecorator';
 import Changeable from '@enact/ui/Changeable';
 import Pure from '@enact/ui/internal/Pure';
 import Toggleable from '@enact/ui/Toggleable';
 import Layout, {Cell} from '@enact/ui/Layout';
+import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
 import React from 'react';
@@ -404,7 +406,7 @@ const InputPopupBase = kind({
 							/> :
 							<InputField
 								{...inputProps}
-								className={css.textField}
+								className={classnames(css.textField, spotlightDefaultClass)}
 								css={css}
 								maxLength={maxLength}
 								minLength={minLength}