+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 - `sandstone/Panels.Header` to always show back button
 - `sandstone/Scroller` scrollbar thumb to not get focus with touch when `focusableScrollbar` is `true` or `byEnter`
 - `sandstone/Scroller` and `sandstone/VirtualList` to show overscroll effect when flicking
-- `sandstone/TabLayout` to collapse tabs by touch flicking
+- `sandstone/TabLayout` to collapse and expand tabs by touch flicking
 
 ### Fixed