+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
-## [unreleased]
+## [1.4.4-touch-alpha] - 2020-11-19
 
 ### Added